fix(logic): split FRR deal-matcher tests under 400-line limit (Refs #2992)
CI repo.logic_test_file_size failed on the combined first-right test file;
move activity/index cases to a supplement file and share frrMatcherTestIndex
via test support.
